Remove unnecessary files in src file & add new components
