--- bookstore-app-2/src/components/BookItem.js
@@ -1,10 +1,0 @@
-const BookItem = ({bookListProp}) => {
-    return (
-        <>
-        <li>{bookListProp.title}</li>
-        <li>{bookListProp.author}</li>
-        </>
-    );
-}
-
-export default BookItem;
