--- bookstore-app-2/src/components/BookList.js
@@ -1,15 +1,0 @@
-import BookItem from './BookItem';
-
-const BookList = ({ booksProps }) => {
-    return (
-        <ul>
-            {
-                booksProps.map((booksArray) => (
-                    <BookItem bookListProp={booksArray} />
-                ))
-            }
-        </ul>
-    );
-}
-
-export default BookList;
