--- bookstore-app-2/src/components/Category.js
@@ -1,7 +1,0 @@
-const Category = () => {
-    return (
-        <h2>Happy Category</h2>
-    );
-}
-
-export default Category;
