--- bookstore-app-2/src/components/InputBook.js
@@ -1,14 +1,0 @@
-const FormInput = () => {
-    return (
-        <form>
-            <input type="text" placeholder="Book title"/>
-            
-            <input type="text" placeholder="author" />
-            
-            <button>Add Book</button>
-            
-        </form>
-    );
-}
-
-export default FormInput;
