+++ src/App.css
@@ -1,0 +1,5 @@
+body {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
