+++ src/App.js
@@ -1,0 +1,21 @@
+import { Route, Routes } from 'react-router-dom';
+import Books from './components/Books';
+import Category from './components/Category';
+import NavBar from './components/Navbar';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <div>
+        <NavBar />
+        <Routes>
+          <Route path="/" element={<Books />} />
+          <Route path="/category" element={<Category />} />
+        </Routes>
+      </div>
+    </div>
+  );
+}
+
+export default App;
