+++ src/components/BookItem.js
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types';
+
+const BookItem = ({ bookListProp }) => (
+  <>
+    <li>{bookListProp.title}</li>
+    <li>{bookListProp.author}</li>
+  </>
+);
+
+export default BookItem;
