+++ src/components/Category.js
@@ -1,0 +1,5 @@
+const Category = () => (
+  <h2>Happy Category</h2>
+);
+
+export default Category;
