+++ src/components/InputBook.js
@@ -1,0 +1,12 @@
+const FormInput = () => (
+  <form>
+    <input type="text" placeholder="Book title" />
+
+    <input type="text" placeholder="author" />
+
+    <button>Add Book</button>
+
+  </form>
+);
+
+export default FormInput;
