+++ src/modules/Navbar.css
@@ -1,0 +1,15 @@
+.navbar-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 80px;
+}
+
+.navbar-links {
+  display: flex;
+  flex-direction: row;
+  gap: 20px;
+  list-style: none;
+  font-size: 25px;
+}
